fix: Sync payment updates to Google Sheets for cross-device consistency
Fixed critical bug where payments made on one device were not visible on other devices.

Root cause:
- Payments were only updating localStorage (device-specific)
- Google Sheets was not being updated with new cuotas_pagadas values
- When syncing from another device, old data from Sheets would overwrite local changes

Changes:
- PaymentForm.tsx: Added updateInSheet() call to update Google Sheets after payment
- google-apps-script-updated.js: Changed parseInt to parseFloat for cuotas_pagadas
  to support decimal values (0.5, 1.75, etc.) from partial payments
- App.tsx: Changed parseInt to parseFloat when syncing cuotas_pagadas from Sheets

Now when you make a payment:
1. Updates locally (localStorage)
2. Updates Google Sheets (Gastos_Pendientes)
3. Registers payment in Sheets (Pagos)
4. Other devices will see the update when they sync

This ensures cross-device consistency and proper data persistence.

+++ App.tsx
@@ -118,7 +118,7 @@ function App() {
           ...p,
           monto: parseFloat(p.monto) || 0,
           num_cuotas: parseInt(p.num_cuotas) || 1,
-          cuotas_pagadas: parseInt(p.cuotas_pagadas) || 0
+          cuotas_pagadas: parseFloat(p.cuotas_pagadas) || 0 // Cambiado a parseFloat para permitir valores decimales
         }));
         savePending(cleanPending);
       }
